% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion_predict.R
\name{diffusion_predict}
\alias{diffusion_predict}
\title{Predict parameters of the drift-diffusion model.}
\usage{
diffusion_predict(df, conditions = 2)
}
\arguments{
\item{df}{The dataset.}

\item{conditions}{The number of conditions.}
}
\value{
NULL.
}
\description{
Predict parameters of the drift-diffusion model.
}
\examples{
diffusion_predict(df, 2)
}
