% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_leaky_models.R
\name{download_leaky_models}
\alias{download_leaky_models}
\title{Load the pretrained LCA models.}
\usage{
download_leaky_models(path = ".")
}
\arguments{
\item{path}{The path to store the downloaded models.}
}
\value{
The pretrained models.
}
\description{
Load the pretrained LCA models.
}
\examples{
download_leaky_models()
}
