% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_diffusion.R
\name{fit_diffusion}
\alias{fit_diffusion}
\title{Predict parameters of the drift-diffusion model.}
\usage{
fit_diffusion(df, conditions = 2)
}
\arguments{
\item{df}{The dataset.}

\item{conditions}{The number of conditions.}
}
\value{
All parameters of the drift-diffusion model (a,t0,...).
}
\description{
Predict parameters of the drift-diffusion model.
}
\examples{
fit_diffusion(df, 2)
}
