% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_leaky.R
\name{fit_leaky}
\alias{fit_leaky}
\title{Predict parameters of the leaky competing accumulator model (LCA).}
\usage{
fit_leaky(df, model)
}
\arguments{
\item{df}{The dataset.}

\item{model}{The loaded Keras model.}
}
\value{
The parameters of the LCA model (a,t0,...).
}
\description{
Predict parameters of the leaky competing accumulator model (LCA).
}
\examples{
download_leaky_models()
input = keras::array_reshape(as.matrix(data.frame(z=rnorm(100),x=0,y=0,z=0)),c(1,100,4))
deepinf = load_leaky(competitors=3)
fit_leaky(input, deepinf)
}
