% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_levy.R
\name{fit_levy}
\alias{fit_levy}
\title{Predict parameters of the Lévy Flight Model.}
\usage{
fit_levy(df, conditions = 2)
}
\arguments{
\item{df}{The dataset.}

\item{conditions}{The number of conditions.}
}
\value{
The parameters of the LCA model (a,t0,...).
}
\description{
Predict parameters of the Lévy Flight Model.
}
\examples{
fit_levy(df,2)
}
