% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leaky3}
\alias{leaky3}
\title{Leaky3.}
\format{A data frame with 100 rows and 4 variables:
\describe{
  \item{a}{price, in US dollars}
  \item{b}{weight of the diamond, in carats}}}
\usage{
leaky3
}
\description{
An examples dataset for the LCA model with 3 competitors.
}
\keyword{datasets}
