% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaky_download.R
\name{leaky_download}
\alias{leaky_download}
\title{Load the pretrained LCA models.}
\usage{
leaky_download(path = ".")
}
\arguments{
\item{path}{The path to store the downloaded models.}
}
\value{
NULL.
}
\description{
Load the pretrained LCA models.
}
\examples{
leaky_download()
}
