% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaky_load.R
\name{leaky_load}
\alias{leaky_load}
\title{Load a pretrained model from the class of leaky competing accumulator models (LCA).}
\usage{
leaky_load(path = ".", competitors = 3)
}
\arguments{
\item{path}{The path where the downloaded models are.}

\item{competitors}{The number of competing processes.}
}
\value{
The pretrained model as a Keras model.
}
\description{
Load a pretrained model from the class of leaky competing accumulator models (LCA).
}
\examples{
leaky_download()
deepinf = leaky_load(competitors=3)
}
