% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaky_predict.R
\name{leaky_predict}
\alias{leaky_predict}
\title{Predict parameters of the leaky competing accumulator model (LCA).}
\usage{
leaky_predict(df, model)
}
\arguments{
\item{df}{The dataset.}

\item{model}{The loaded Keras model.}
}
\value{
The parameters of the LCA model (a,t0,...).
}
\description{
Predict parameters of the leaky competing accumulator model (LCA).
}
\examples{
leaky_download()
input = keras::array_reshape(as.matrix(data.frame(z=rnorm(100),x=0,y=0,z=0)),c(1,100,4))
deepinf = leaky_load(competitors=3)
leaky_predict(input, deepinf)
}
