% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaky_sim.R
\name{leaky_sim}
\alias{leaky_sim}
\title{Simulate data from the class of leaky competing accumulator models (LCA).}
\usage{
leaky_sim(trials)
}
\arguments{
\item{trials}{The number of trials}
}
\value{
The simulated dataset and the actual parameters used for simulation (list).
}
\description{
Simulate data from the class of leaky competing accumulator models (LCA).
}
\examples{
out_list = leaky_sim(1000)
}
