% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levy_predict.R
\name{levy_predict}
\alias{levy_predict}
\title{Predict parameters of the Lévy Flight Model.}
\usage{
levy_predict(df, conditions = 2)
}
\arguments{
\item{df}{The dataset.}

\item{conditions}{The number of conditions.}
}
\value{
The parameters of the LCA model (a,t0,...).
}
\description{
Predict parameters of the Lévy Flight Model.
}
\examples{
levy_predict(df,2)
}
