% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_leaky.R
\name{load_leaky}
\alias{load_leaky}
\title{Load a pretrained model from the class of leaky competing accumulator models (LCA).}
\usage{
load_leaky(path = ".", competitors = 3)
}
\arguments{
\item{path}{The path where the downloaded models are.}

\item{competitors}{The number of competing processes.}
}
\value{
The pretrained model as a Keras model.
}
\description{
Load a pretrained model from the class of leaky competing accumulator models (LCA).
}
\examples{
download_leaky_models()
deepinf = load_leaky(competitors=3)
}
