% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare a dataset for prediction.}
\usage{
prepare_data(df)
}
\arguments{
\item{df}{The dataset to be prepared.}
}
\value{
The preprocessed dataset.
}
\description{
Prepare a dataset for prediction.
}
\examples{
df = leaky_sim(100)[[1]]
input = prepare_data(df)
}
