% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffusion_sim.R
\name{diffusion_sim}
\alias{diffusion_sim}
\title{Simulate data from the diffusion model.}
\usage{
diffusion_sim(n, a, v, t0, z = 0.5 * a, d = 0, sz = 0, sv = 0,
  st0 = 0, s = 1, precision = 3, stop_on_error = TRUE)
}
\arguments{
\item{trials}{The number of trials}
}
\value{
The simulated dataset.
}
\description{
Simulate data from the diffusion model.
}
\examples{
df = diffusion_sim(500, a = 1, v = 2, t0 = 0.5)
}
